Experimental changes to use ratio for ts-wordlength instead of absolute values.

+++ src/main/java/org/briljantframework/mimir/data/timeseries/SaxOptions.java
@@ -11,7 +11,7 @@ public class SaxOptions {
     private static Map<Integer, double[]> breakpoints = new HashMap<Integer, double[]>();
     private static double[][] distTable;
     private static int alphabetSize;
-    private static int tsWordLength;
+    private static double tsWordLength;
     private static double lowerWordLength;
     private static double upperWordLength;
 
@@ -24,11 +24,11 @@ public class SaxOptions {
         SaxOptions.alphabetSize = alphabet_size;
     }
 
-    public static int getTsWordLength() {
+    public static double getTsWordLength() {
         return tsWordLength;
     }
 
-    public static void setTsWordLength(int word_length) {
+    public static void setTsWordLength(double word_length) {
         SaxOptions.tsWordLength = word_length;
     }
